requirements.txt
# Core Framework
Flask>=3.0.0
Flask-Cors>=4.0.0
Werkzeug>=3.0.0

# API & Async
requests>=2.31.0
aiohttp>=3.9.1

# AI/ML Services
openai>=1.3.0
google-generativeai>=0.3.0

# Voice Processing (Optional - install separately if needed)
# SpeechRecognition>=3.10.0
# pyttsx3>=2.90
# PyAudio requires system dependencies - install with: pipwin install pyaudio

# Environment & Configuration
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2

# Production Server
gunicorn>=21.0.0

# Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type Checking (Optional)
typing-extensions>=4.8.0
